requirements_win.txt
ansimarkup==1.4.0
app-state==0.5
better-exceptions-fork==0.2.1.post6
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
docutils==0.14
getinstance==0.4
h11==0.8.1
h2==3.1.0
hpack==3.0.0
httpx==0.6.8
hyperframe==5.2.0
idna==2.8
lockorator==0.1
loguru==0.2.5
minidjango==2.2.3.dev20190701035504
plyer==1.4.0
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-async==0.6.2
rfc3986==1.3.2
six==1.12.0
sniffio==1.1.0
sqlparse==0.3.0
urllib3==1.25.3

pypiwin32 
kivy_deps.sdl2==0.1.22 
kivy_deps.glew==0.1.12
